+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Bug fixes**
 
 - Made `EuiIconTip` screen reader accessible ([#534](https://github.com/elastic/eui/pull/534))
+- Fixed a sorting issue in `EuiInMemoryTable` ([#453](https://github.com/elastic/eui/pull/453))
 
 # [`0.0.30`](https://github.com/elastic/eui/tree/v0.0.30)